Benchmarking/requirements.txt
numpy
soundfile
torchaudio
torch
onnxruntime
openai-whisper
pyyaml
tqdm
